Provide a simpler way of sniffing the right platform client to use, and isolate use of EmptyApp
